Basic performance testing based on bogomips style loop.

+++ core/src/cpp/include/mc64k.hpp
@@ -14,12 +14,12 @@
 #ifndef __MC64K_HPP__
 #  define __MC64K_HPP__
 
-typedef unsigned long long uint64;
-typedef signed long long   int64;
-typedef unsigned long      uint32;
-typedef signed long        int32;
-typedef unsigned short     uint16;
-typedef signed short       int16;
+typedef unsigned long int  uint64;
+typedef signed long int    int64;
+typedef unsigned int       uint32;
+typedef signed int         int32;
+typedef unsigned short int uint16;
+typedef signed short int   int16;
 typedef unsigned char      uint8;
 typedef signed char        int8;
 typedef double             float64;
